Moved Functions to Separate File
